+++ src/bbs2gh/Commands/GenerateScriptCommand.cs
@@ -23,7 +23,7 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         AddOption(SshUser);
         AddOption(SshPrivateKey);
         AddOption(SshPort);
-        AddOption(SshHost);
+        AddOption(ArchiveDownloadHost);
         AddOption(SmbUser);
         AddOption(SmbDomain);
         AddOption(Output);
@@ -59,9 +59,9 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         description: "Bitbucket server's shared home directory. Defaults to \"/var/atlassian/application-data/bitbucket/shared\" if downloading the archive from a server using SSH " +
                      "and \"c$\\atlassian\\applicationdata\\bitbucket\\shared\" if downloading using SMB.");
 
-    public Option<string> SshHost { get; } = new(
-        name: "--ssh-host",
-        description: "The host to use to connect to the Bitbucket Server via SSH. Defaults to the host from the Bitbucket Server URL (`--bbs-server-url`).");
+    public Option<string> ArchiveDownloadHost { get; } = new(
+        name: "--archive-download-host",
+        description: "The host to use to connect to the Bitbucket Server via SSH or SMB. Defaults to the host from the Bitbucket Server URL (--bbs-server-url).");
 
     public Option<string> SshUser { get; } = new(
         name: "--ssh-user",
@@ -153,7 +153,7 @@ public class GenerateScriptCommandArgs
     public string BbsPassword { get; set; }
     public string BbsProjectKey { get; set; }
     public string BbsSharedHome { get; set; }
-    public string SshHost { get; set; }
+    public string ArchiveDownloadHost { get; set; }
     public string SshUser { get; set; }
     public string SshPrivateKey { get; set; }
     public int SshPort { get; set; }