+++ RELEASENOTES.md
@@ -1,3 +1,3 @@
 - Adds retry logic during GHES archive generation in cases of transient failure
 - Added log output linking to migration log URL after migration completes
-- Add support for specifying `--ssh-host` with `gh bbs2gh migrate-repo` and `gh bbs2gh generate-script`, rather than taking the host from the `--bbs-server-url`
+- Add support for specifying `--archive-download-host` with `gh bbs2gh migrate-repo` and `gh bbs2gh generate-script`, rather than taking the host from the `--bbs-server-url`